+++ src/layout/Customer/desktop/CustomerLayout.jsx
@@ -1,6 +1,6 @@
-import Header from "../../components/Customer/organisms/header/Header";
+import Header from "../../../components/Customer/organisms/desktop/header/Header";
 import { Outlet } from "react-router-dom";
-import Footer from "../../components/Customer/organisms/footer/Footer";
+import Footer from "../../../components/Customer/organisms/desktop/footer/Footer";
 import "./CustomerLayout.css";
